// src/api-client/models/order-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spring-Adam-Store
 * Api Service
 *
 * The version of the OpenAPI document: v1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AddressResponse } from './address-response';
// May contain unused imports in some cases
// @ts-ignore
import type { OrderItemResponse } from './order-item-response';

/**
 * 
 * @export
 * @interface OrderResponse
 */
export interface OrderResponse {
    /**
     * 
     * @type {number}
     * @memberof OrderResponse
     */
    'id'?: number;
    /**
     * 
     * @type {string}
     * @memberof OrderResponse
     */
    'orderDate'?: string;
    /**
     * 
     * @type {number}
     * @memberof OrderResponse
     */
    'discountAmount'?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderResponse
     */
    'totalPrice'?: number;
    /**
     * 
     * @type {string}
     * @memberof OrderResponse
     */
    'orderStatus'?: OrderResponseOrderStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof OrderResponse
     */
    'customerName'?: string;
    /**
     * 
     * @type {AddressResponse}
     * @memberof OrderResponse
     */
    'address'?: AddressResponse;
    /**
     * 
     * @type {Array<OrderItemResponse>}
     * @memberof OrderResponse
     */
    'orderItems'?: Array<OrderItemResponse>;
}

export const OrderResponseOrderStatusEnum = {
    Pending: 'PENDING',
    Processing: 'PROCESSING',
    Shipped: 'SHIPPED',
    Delivered: 'DELIVERED',
    Cancelled: 'CANCELLED'
} as const;

export type OrderResponseOrderStatusEnum = typeof OrderResponseOrderStatusEnum[keyof typeof OrderResponseOrderStatusEnum];